feat: 일단 구현 수정

+++ src/main/java/com/team2/service/CartService.java
@@ -1,6 +1,7 @@
 package com.team2.service;
 
 import com.team2.dto.cart.CartDTO;
+import com.team2.dto.cart.CartResponse;
 import com.team2.mapper.CartMapper;
 import com.team2.model.Cart;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,8 +17,8 @@ public class CartService {
     CartMapper cartMapper;
 
     // 회원의 장바구니 리스트 조회
-    public List<CartDTO> getCartList(CartDTO cartDTO) {
-        List<CartDTO> result = cartMapper.getCartList(cartDTO);
+    public List<CartResponse> getCartList(CartDTO cartDTO) {
+        List<CartResponse> result = cartMapper.getCartList(cartDTO.getCustomerId());
         return result;
     }
 
@@ -36,7 +37,9 @@ public class CartService {
 
         if (existing != null) {
             throw new IllegalArgumentException("장바구니에 동일 상품이 존재합니다.");
-        } else cartMapper.addCart(cartDTO);
+        } else {
+            cartMapper.addCart(cartDTO);
+        }
     }
 
     // 장바구니 인원 수정
@@ -47,24 +50,40 @@ public class CartService {
         }
 
         // 장바구니 총 인원이 룸 최대 인원보다 많을 경우 검사
-        int total = cartDTO.getAdultCount() + cartDTO.getChildCount();
-        int capacity = PagingConditionDTO.getTotalPepleCnt();
-        //int capacity = roomService.getRoomCapacity(cartDTO.getRoomId()); ??
-        if (total > capacity)
-        throw new IllegalArgumentException("선택하신 인원(" + total + "명)이 객실 수용인원(" + capacity + "명)을 초과합니다.");{
-        }
+//        int total = cartDTO.getAdultCount() + cartDTO.getChildCount();
+//        int capacity = PagingConditionDTO.getTotalPepleCnt();
+//        //int capacity = roomService.getRoomCapacity(cartDTO.getRoomId()); ??
+//        if (total > capacity)
+//        throw new IllegalArgumentException("선택하신 인원(" + total + "명)이 객실 수용인원(" + capacity + "명)을 초과합니다.");{
+//        }
 
         Cart cart = new Cart();
+        cart.setDesiredCheckInAt(cartDTO.getDesiredCheckInAt());
+        cart.setDesiredCheckOutAt(cartDTO.getDesiredCheckOutAt());
+
         if (!cart.isValidDateRange()) {
             throw new IllegalArgumentException("체크아웃 날짜는 체크인 날짜 이후여야 합니다.");
         }
-        cartMapper.addCart(cartDTO);
+//        cartMapper.addCart(cartDTO);
 
         return cartMapper.updatePeopleCnt(cartDTO);
     }
 
     // 장바구니 삭제
-    public void deleteCart(int cartId) {
-        cartMapper.deleteCart(cartId);
+    public void deleteCart(int cartId, int customerId) {
+
+        // 동일 회원 + 동일 객실 중복 방지 -> 과제 평가사항에 있음
+        CartDTO existing = cartMapper.findByCustomerIdAndCartId(
+                customerId,
+                cartId
+        );
+
+        if (existing == null) {
+            throw new IllegalArgumentException("장바구니에 해당하는 정보가 없습니다.");
+        } else {
+            cartMapper.deleteCart(cartId);
+        }
+
+
     }
 }
